refactor: moved config to one place
