+++ docs/ci-cd-setup.md
@@ -57,7 +57,6 @@ The code for the Master branch is:
 nvm install 8
 npm i -g npm@latest
 cd backend
-sed -i -e 's/localhost/mongo/g' config/config.tsx
 npm ci
 docker build -t amosconversationaltax/conversational-tax .
 docker push amosconversationaltax/conversational-tax
@@ -70,7 +69,6 @@ The code for the Develop branch is:
 nvm install 8
 npm i -g npm@latest
 cd backend
-sed -i -e 's/localhost/mongo/g' config/config.tsx
 npm ci
 docker build -t amosconversationaltax/conversational-tax-dev .
 docker push amosconversationaltax/conversational-tax-dev
@@ -86,7 +84,6 @@ cd ../frontend
 npm ci
 npm i -g exp
 exp login -u $USERNAME -p $PASSWORD
-sed -i -e 's/localhost:3000/$URL:$PORT/g' config/config.tsx
 sed -i -e 's/conversational-tax/conversational-tax-$BRANCH/g' app.json
 sed -i -e 's/Conversational Tax/Conversational Tax ($BRANCH)/g' app.json
 npm run publish
@@ -116,7 +113,7 @@ services:
         links:
             - mongo
         volumes:
-            - ~/amos_files/dialogflowKey.json:/usr/src/app/dialogflowKey.json
+            - ~/amos_files/config/:/usr/src/app/dist/config/
     mongo:
         restart: always
         image: mongo
@@ -144,7 +141,7 @@ services:
         links:
             - mongo
         volumes:
-            - ~/amos_files/dialogflowKey.json:/usr/src/app/dialogflowKey.json
+            - ~/amos_files/config/:/usr/src/app/dist/config/
     mongo:
         restart: always
         image: mongo
